requirements.txt
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==37.0.1
ecdsa==0.17.0
Flask==2.1.1
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
google-auth==2.6.6
google-auth-oauthlib==0.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jose==1.0.0
MarkupSafe==2.1.1
oauthlib==3.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.3.1
pyOpenSSL==22.0.0
python-jose==3.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.35
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
